requirements.txt
streamlit
streamlit_option_menu
pillow
plotly
